+++ example/gifplayer/gifplayer.c
@@ -330,7 +330,8 @@ const char* play() {
     wuffs_base__pixel_config__set(
         &ic.pixcfg,
         wuffs_base__make_pixel_format(WUFFS_BASE__PIXEL_FORMAT__BGRA_PREMUL),
-        WUFFS_BASE__PIXEL_SUBSAMPLING__NONE, width, height);
+        wuffs_base__make_pixel_subsampling(WUFFS_BASE__PIXEL_SUBSAMPLING__NONE),
+        width, height);
 
     const char* msg = allocate(&dec, &ic);
     if (msg) {