+++ doc/changelog.md
@@ -5,6 +5,7 @@
 
 - Renamed warnings to notes.
 - Made `wuffs_base__pixel_format` a struct.
+- Made `wuffs_base__pixel_subsampling` a struct.
 - Made `wuffs_base__status` a struct.